Add support for multiple OpenID Connect providers
The login.tpl, AuthController.php, baseTemplate.tpl, user.php, and various other files were modified to add support for multiple OpenID Connect providers. An "OR" section was introduced in the login page, allowing users to choose an OpenID provider for authentication. Functionality for logging in with specific providers is handled in a new OpenID controller, with each provider identified by a unique key. This improves user flexibility and security during the login process.